refactor(app): Phase 2 - pages fetch data, remove loaders
- Move data fetching to page level (Home, Projects, Blog)
- Pages fetch from lib/data, pass props to presentation components
- Remove FeaturedProjectsLoader, ProjectListLoader, BlogListLoader
- Add (ide)/loading.tsx for route segment loading state
- IdeLayoutWithCommits uses lib/data
- Blog opengraph uses lib/data
- Update code-content welcome sample to match new pattern
