+++ apps/web/src/app/[locale]/(ide)/loading.tsx
@@ -1,0 +1,5 @@
+import { PageContentSkeleton } from "@/components/ide/layout/page-content-skeleton";
+
+export default function IdeLoading() {
+  return <PageContentSkeleton />;
+}
