+++ apps/web/src/app/[locale]/(ide)/projects/page.tsx
@@ -1,10 +1,11 @@
 import { getTranslations, setRequestLocale } from "next-intl/server";
 import { CodeView } from "@/components/ide/editor/code-view";
 import { EditorContent } from "@/components/ide/editor/editor-content";
-import { ProjectListLoader } from "@/components/projects/project-list-loader";
+import { ProjectList } from "@/components/projects/project-list";
 import { ProjectsHeader } from "@/components/projects/projects-header";
 import { PageContent } from "@/components/shared/page-content";
 import { projectsCode } from "@/consts/code-content";
+import { getGitHubRepos } from "@/lib/data";
 import { createMetadata } from "@/lib/metadata";
 
 export async function generateMetadata({
@@ -28,12 +29,14 @@ export default async function Projects({ params }: Props) {
   const { locale } = await params;
   setRequestLocale(locale);
 
+  const repos = await getGitHubRepos();
+
   return (
     <EditorContent
       preview={
         <PageContent>
           <ProjectsHeader />
-          <ProjectListLoader />
+          <ProjectList repos={repos} />
         </PageContent>
       }
       source={<CodeView code={projectsCode} lang="typescript" />}
