--- apps/web/src/components/projects/project-list-loader.tsx
@@ -1,7 +1,0 @@
-import { ProjectList } from "@/components/projects/project-list";
-import { getGitHubRepos } from "@/lib/github";
-
-export async function ProjectListLoader() {
-  const repos = await getGitHubRepos();
-  return <ProjectList repos={repos} />;
-}
